Game Logic implementation and some working done atleast
BADLY STUCK AT Delays not working and then realized that even the static variables inside the function calls belong to the whole class and no the object

HAve to remake Utility class again

+++ Transform.hpp
@@ -13,7 +13,7 @@ public:
     int zrot;
 
 //=========int scale===========
-    int x_sc,y_sc;
+    float x_sc,y_sc;
     Transform(int _x , int _y );
 
     SDL_Rect* ToScreenPosition();
